Bug/Issue#13/InvalidAppVeyorInstallCommand
Initially AppVeyor did not support NUnit3.0, as a result the appVeyor.yml file included a installation step which downloads NUnit 3.0 beta and adds it to the system path. This step is now invalid as AppVeyor has updated their service to use NUnit 3.0 for test executions and this custom installation step is failing to run.

As a fix, this custom installation step has been removed.